requirements.txt
# Home Brain — Python Dependencies
# Requires Python 3.9+
# ffmpeg must be installed separately as a system package:
#   Ubuntu/Debian: sudo apt install ffmpeg
#   macOS:         brew install ffmpeg

# ── Phase 1 — Surveillance Domain ────────────────────
python-dotenv>=1.0.0

# ── Phase 2 — Intelligence Domain ────────────────────────
anthropic>=0.25.0
faster-whisper>=1.0.0

# ── Phase 3 — Memory Domain (uncomment when ready) ───
# python-telegram-bot>=20.0
# boto3>=1.34.0
